+++ tests/testthat/test-vlsig.R
@@ -33,19 +33,43 @@ test_that("vl_generators works", {
 
 test_that("vlsig works", {
   major_triad <- c(0, 4, 7)
-  expect_equal(vlsig(major_triad)$"vl", c(0, 1, 2))
-  expect_equal(vlsig(major_triad)$"tn", 5)
-  expect_equal(vlsig(major_triad)$"rotation", 1)
+  maj_triad_res <- matrix(c(0, 1, 2, 1, 0, 2), nrow=2, byrow=TRUE)
+  expect_equal(vlsig(major_triad, index=1)$"vl", c(0, 1, 2))
+  expect_equal(vlsig(major_triad, index=1)$"tn", 5)
+  expect_equal(vlsig(major_triad, index=1)$"rotation", 1)
+  expect_equal(vlsig(major_triad, index=NULL), maj_triad_res)
 
-  expect_equal(vlsig(major_triad, 2)$"vl", c(1, 0, 2))
-  expect_equal(vlsig(major_triad, 2)$"tn", 9)
-  expect_equal(vlsig(major_triad, 2)$"rotation", 2)
+  expect_equal(vlsig(major_triad, index=2)$"vl", c(1, 0, 2))
+  expect_equal(vlsig(major_triad, index=2)$"tn", 9)
+  expect_equal(vlsig(major_triad, index=2)$"rotation", 2)
 
-  expect_equal(vlsig(c(0, 6, 11), edo=19)$"vl", c(0, 2, 3))
+  expect_equal(vlsig(c(0, 6, 11), index=1, edo=19)$"vl", c(0, 2, 3))
 
   gh <- c(0, 2, 4, 5, 7, 9)
-  expect_equal(vlsig(gh, 1)$"vl", c(0, 1, 1, 2, 1, 1))
+  expect_equal(vlsig(gh, index=1)$"vl", c(0, 1, 1, 2, 1, 1))
 
-  expect_error(vlsig(c(0, 2, 4, 5, 7, 9, 11), 2))
-  expect_error(vlsig(major_triad, 0))
+  expect_error(vlsig(c(0, 2, 4, 5, 7, 9, 11), index=2))
+  expect_error(vlsig(major_triad, index=0))
+})
+
+test_that("inter_vlsig works", {
+  jdia_matrix <- matrix(c(0, 0, 0, .92, 0, 0, 0,
+                   .71, .71, 0, .71, .71, 0, 0,
+                   .22, 0, .22, .22, .22, .22, .22),
+                   nrow=3, byrow=TRUE)
+  expect_equal(inter_vlsig(j(dia)), jdia_matrix)
+  expect_equal(inter_vlsig(j(dia), index=1)$tni, j(7))
+  expect_equal(inter_vlsig(j(dia), index=1)$rotation, 0)
+
+  triad_matrix <- matrix(c(0, 0, 2, 1, 0, 1, 0, 1, 1), 
+                         nrow=3, byrow=TRUE)
+  expect_equal(inter_vlsig(c(0, 4, 7)), triad_matrix)
+
+  hexachord <- c(0, 2, 7, 12, 16, 17)
+  h_res1 <- list(vl=c(0, 2, 2, 2, 0, 6), tni=16, rotation=5)
+  h_res2 <- list(vl=c(-2, 0, 0, 0, -2, 4), tni=14, rotation=5)
+  expect_equal(inter_vlsig(hexachord, index=1, type="ascending", edo=24),
+               h_res1)
+  expect_equal(inter_vlsig(hexachord, index=1, type="commontone", edo=24),
+               h_res2)  
 })